+++ tests/Neo.Plugins.L2Bridge.UnitTests/Usings.cs
@@ -1,5 +1,6 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Neo;
 global using Neo.L2;
+global using Neo.L2.Bridge;
 global using Neo.L2.Telemetry;
 global using Neo.Plugins.L2;